Commented out all "pause(eps)" statements
This is a bugfix for a problem identified in solvers/getCobraSolverParams.m.
Profiling results indicated that pause(eps) calls accounted for a large majority
of the time spent in this function (and in the program that called it). Further
discussion with Ronan Fleming confirmed that this behavior is not intentional,
and the pause statement should be removed. In this commit I remove it from all
(non-doc) files.