Changed file extension to avoid mcc parse error
MCC parses all included .m files to search for dependencies. The renamed file is
has a .m extension, but is NOT a MATLAB script. As a result, MCC fails to parse
the file with numerious syntax errors. The fix is to rename the file itself and
update it's name in the one function that makes use of it.