Removing a for loop to speed up gpSampler
The for loop here with the vectors whose sizes are not predesignated makes the computation very slow when the size goes up. For models of size ~ 10000, this can take up to 10 seconds while the proposed code takes >0.01. They produce exactly the same matrix Anew, Bnew and vector Cnew, Dnew and the check of input 'csense'.