createTissueSpecificModel bugfix for geneids containing containing specific letters.
The extractGPRs function in createTissueSpecificModel would not work if
geneids which contained containing any of the following letters: "adnor"
as the strtok function splits strings upon encountering any delimiter.
From the documentation for strtok:

If the delimiter input specifies more than one character, MATLAB® treats
each character as a separate delimiter; it does not treat the multiple
characters as a delimiting string.

Replacing and/or with &/| lets strtok parsing occur normally.