Update convertToIrreversible.m
When I wanted to sample my model I needed to first make it irreversible. However after applying convertToIrreversible function (which worked OK) I couldn't use the obtained model_Irrev for the sampleCbModel.m as it didn't have the grRules field in it. I updated the convertToIrreversible.m function to added that field to the newly created model.