Fixed to prior to 2016a
Before 2016a (or maybe b, this feature is undocumented), the TC.xml file in the matlab root was used for tab completion. After that functionSignatures.json files are required. The added lines check the matlab version and whether the user has write access, and if, it adds the tab completion.
If not, nothing happens, and no tab completion is available