Protect against some (but not all) unbounded fluxes
The quality of the solution of the first FBA is checked via tempsol.stat, but the quality of the solution of the maximization/minimization  of the sum of all fluxes is not checked.  Therefore, if some but not all fluxes are unbounded the function fails with an unspecific error on line 258. 

Change inspired by: https://groups.google.com/forum/#!topic/cobra-toolbox/CJ8YAfueV58